Small changes to state_managment_template, improved the functions in mobx example and added a MobX reaction for freight

+++ flutter_state_management_mobx/lib/global/functions/app_dialogs.dart
@@ -5,7 +5,7 @@ import 'package:state_management_example/global/styles/app_fonts.dart';
 class AppDialog {
   static String addOnClick({TextEditingController textEditingController}) {
     print('Add quantity here!');
-    double currentValue = double.parse(textEditingController.text);
+    int currentValue = int.parse(textEditingController.text);
     if (currentValue < 999) {
       currentValue++;
     }
@@ -14,7 +14,7 @@ class AppDialog {
 
   static String subtrackOnClick({TextEditingController textEditingController}) {
     print('Subtract quantity here!');
-    double currentValue = double.parse(textEditingController.text);
+    int currentValue = int.parse(textEditingController.text);
     if (currentValue > 0) {
       currentValue--;
     }
@@ -26,23 +26,23 @@ class AppDialog {
     return quantity;
   }
 
-  static void cancelOnClick({BuildContext context, double quantity}) {
+  static void cancelOnClick({BuildContext context, int quantity}) {
     print('Cancel change here!');
     Navigator.of(context).pop(quantity);
   }
 
-  static void confirmOnClick({BuildContext context, double newQuantity}) {
+  static void confirmOnClick({BuildContext context, int newQuantity}) {
     print('Confirm quantity here!');
     Navigator.of(context).pop(newQuantity);
   }
 
-  static Future<double> quantityDialog(
+  static Future<int> quantityDialog(
       {@required BuildContext context,
       @required String title,
-      @required double quantity}) async {
+      @required int quantity}) async {
     TextEditingController textEditingController = TextEditingController();
     textEditingController.text = quantity.toString();
-    final newQuantity = await showDialog<double>(
+    final newQuantity = await showDialog<int>(
       context: context,
       barrierDismissible: false,
       builder: (BuildContext context) {
@@ -127,8 +127,7 @@ class AppDialog {
                       splashColor: Colors.grey,
                       onPressed: () => AppDialog.confirmOnClick(
                           context: context,
-                          newQuantity:
-                              double.parse(textEditingController.text)),
+                          newQuantity: int.parse(textEditingController.text)),
                       child: Container(
                           child: Text(
                         'Confirm',
